Make search case-insensitive
JQuery's contains() filter function is case sensitive. It seems more
natural to have search be case insensitive. This solution was taken
from a stack overflow question:

https://stackoverflow.com/questions/35480973/jquery-filtering-case-insensitive

If anyone feels that this code is inappropriate code to use, please say so,
and I'll try to work out a different solution (this works pretty well
though)