Make search case-insensitive
JQuery's contains() filter function is case sensitive. It seems more
natural to have search be case insensitive. This solution was taken
from an example on jquery's Sizzle wiki page:

https://github.com/jquery/sizzle/wiki#-backwards-compatible-plugins-for-pseudos-with-arguments

If anyone feels that this code is inappropriate code to use, please say so,
and I'll try to work out a different solution (this works pretty well
though). I fell like this usage is acceptable, as it is fully open source, and
well cited.